requirements.txt
numpy
scikit-learn
streamlit
Pillow 